frontend/requirements.txt
streamlit>=1.43.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.7.0
